Version 4
I edited the text on the home page giving the titles <h1> and the
subtitle <p>

I can’t get the images on the right of the info page to move left more.
They need a space of 20px I think from the images on the left.

I also need to replace some of the images on the home page and info
page.