Fixing info position
There was not symmetry in the two columns and the natural ol/li
formatting was pushing the right column over the edge.

I placed the left column in a list to match the general format of the
site. It is a list with only one element but it is used for sake of a
standard approach. Also, it can be expanded on (more elements added to
left) if needed in the future.

It should be noted that I told Amy this was a bad idea. She was right.